+++ solutions/swb-reference/integration-tests/tests/isolated/costCenter/list.test.ts
@@ -108,7 +108,7 @@ describe('List Cost Center negative tests', () => {
             e,
             new HttpError(400, {
               error: 'Bad Request',
-              message: `Invalid Pagination Token: ${queryParams.paginationToken}`
+              message: `Invalid Pagination Token`
             })
           );
         }