+++ workbench-core/example/app/src/services/authenticationService.ts
@@ -12,8 +12,11 @@ import {
 const cognitoPluginOptions: CognitoAuthenticationPluginOptions = {
   cognitoDomain: process.env.COGNITO_DOMAIN!,
   userPoolId: process.env.USER_POOL_ID!,
-  clientId: process.env.CLIENT_ID!,
-  clientSecret: process.env.CLIENT_SECRET!
+  webUiClient: {
+    clientId: process.env.WEB_UI_CLIENT_ID!,
+    clientSecret: process.env.WEB_UI_CLIENT_SECRET!
+  },
+  allowedClientIds: [process.env.PROGRAMMATIC_ACCESS_CLIENT_ID!]
 };
 
 export const authenticationService: AuthenticationService = new AuthenticationService(