+++ workbench-core/infrastructure/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log - @aws/workbench-core-infrastructure
 
-This log was last generated on Wed, 07 Jun 2023 21:47:22 GMT and should not be manually modified.
+## 1.0.1
+Wed, 16 Aug 2023
+
+### Patches
+* Allow multiple App clientIds for Cognito
 
 ## 1.0.0
 Wed, 07 Jun 2023 21:47:22 GMT